Initial support for authentication state caching via cookies
Added AuthRemoteCookie directive.

  AuthRemoteCookie hrapp_auth_cookie 150

Advises mod_auth_remote to create a cookie called hrapp_auth_cookie that expires in 150 secs.
Once a user logs in successfully, for the next 150 secs he can log in without mod_auth_remote
performing an actual login on the backend.

The cookie (handling code) is reasonably secure. The cookie contains the username, timestamp of
authenticated login and an MD5 of the username,timestamp and a secret random string.

When a mismatch b/w r->remote_user and cookie user occurs or a mismatch between the signature in the cookie
and the signature regenerated at the server occurs or the cookie expires the user is authenticated against
the actual backend authentication system.